Added functionality to most functions
For scenes to play correctly the Garry's Mod addon Portal 2 Voice Lines Fix (https://steamcommunity.com/sharedfiles/filedetails/?id=2691231352) is required for many of the scenes and sounds as of right now because it turns out Portal 2 sounds are incompatible with Garry's Mod because they are from a different engine that is not source 2013. Hence, everything needs to be converted, also if Wheatley's animations seem broken that is most likely because the game normally would use ent_fire to set his animations, which we have not figured out how to do properly yet.